Implemented filters, fixed second Activity
Not sure it's because of emulator, the files are not able to be saved on
the SD card. Lucas, would you please want to test on your phone again?
Note that the images should be stored in the SD card.  The last thing
needs to be fixed is the UPLOAD--still cannot upload large images.